+++ Bellgym_record/testid/2023-12-19.txt
@@ -1,0 +1,6 @@
+이름 : 정회영
+입력날짜는 : 2023년 12월 19일 입니다.
+오늘의 prepare는 : test
+오늘의 mainwork은 : test
+오늘의 wod는 : test
+오늘의 buildup은 test
