+++ Bellgym_record김민오/김민오-2023-12-19.txt
@@ -1,0 +1,6 @@
+이름 : 김민오
+입력날짜는 : 2023년 12월 19일 입니다.
+오늘의 prepare는 : band
+오늘의 mainwork은 : bench
+오늘의 wod는 : bellgymwod3
+오늘의 buildup은 hip band
