+++ constant.py
@@ -13,3 +13,4 @@ UPDATE_USER = 7
 DELETE_USER = 8
 
 FINISH_WORKOUT = 9
+
